Issue #3 - Integrating with Firebase
The necessary API keys to use this integration with Firebase is not
being commited. To execute correctly, get the keys from the Firebase
console.

The specification of which database is being used is on the config
file of the project.

The Sessions.vue doesn't know anymore how to get the notes. It
just receives the notes and refresh the screen.

It fixes #3.